Refactor the noise_maker impl into a cpp file -a
