+++ src/CMakeLists.txt
@@ -10,6 +10,7 @@ find_package(fmt CONFIG REQUIRED)
 
 add_executable(beluga
     beluga.m.cpp
+    noise_maker.cpp
 )
 
 target_link_libraries(beluga PRIVATE
