+++ src/audio_buffer.h
